% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lrr_maf.R
\name{plot_lrr_maf}
\alias{plot_lrr_maf}
\title{CNV raw data plot}
\usage{
plot_lrr_maf(chr, cnv.start, cnv.end, region = 5, final.rep,
  window = 100, step = 50)
}
\arguments{
\item{chr}{Chromosome where the putative CNV is present (e.g "1", not "chr1")}

\item{cnv.start}{Start position of the CNV}

\item{cnv.end}{End position of the CNV}

\item{region}{Size of the region to be plotted on each side of the CNV}

\item{final.rep}{Data frame in Illumina Final report format}

\item{window}{Sliding window size (in SNPs)}

\item{step}{Sliding window step (in SNPs)}
}
\value{
pl
}
\description{
\code{lrr_plot} create scatter plots of LRR and MAF 
in a selected region
}
\details{
This function can be used to inspect the raw data (LRR and MAF)
of putative CNV calls. It also calculate the mean LRR in a 
sliding window with default size of 100 SNPs and default step of
50 SNPs. A bar in each plot indicate the position of the call.
\newline
For convenience consider to select the data from the entire FinalREport
outside R, e.g. using awk. As an example, \code{awk -F"\t" '{if ($1 == "sampleID" && $3 == 1) 
print $0}' FinalReport.txt > chr1_sampleID.tsv} will select the SNPs 
in chromosome 1 of a desired sample and save it as a separate file 
that can be more easily loaded in R.
\newline 
In any case \code{final.rep} must be a data-frame containing the data 
of one single sample and the following columns: "Chr", "Position",
"Log.R.Ratio", "B.allele.Freq". Order is not relevant.
}
\author{
Simone Montalbano simone.montalbano@protonmail.com
}
